fix mike price

+++ ots-web/src/main/java/com/mk/ots/roomsale/service/RoomSaleService.java
@@ -55,5 +55,5 @@ public interface RoomSaleService {
 	public Boolean checkRoomSaleWithOldRoomType(TRoomSaleConfig bean);
 	public List<RoomSaleToIndexDto> getUpdateIndexList(TRoomSaleConfig bean);
 	public List<Map<String, Object>> queryRoomByHotelAndRoomType(String hotelId, String roomTypeId) throws Exception;
-	public BigDecimal getHotelMinPromoPrice(Integer hotelid);
+	public Double getHotelMinPromoPrice(Integer hotelid);
 }
